Refactor set node type attributes to new module
